Extract comparer.js
