+++ lib/replication/remote.js
@@ -2,13 +2,10 @@
 // node lib/replication/remote.js http://admin:admin@author.com/content/tenant http://admin:admin@publish.com/content/tenant
 
 const axios = require('axios')
-const lodash = require('lodash')
-
-const SKIP_PROPS = ['jcr:uuid', 'jcr:lastModified', 'jcr:lastModifiedBy']
-
 const consts = require('./constants')
 const network = require('./network')
 const perapi = require('./perapi')
+const comparer = require('./comparer')
 
 const args = process.argv.slice(2)
 let index = 0
@@ -49,58 +46,15 @@ function compareContent(files) {
     }
 
     Promise.all(fetchPromises).then(contents => {
-        const changedFiles = []
-        const changedJsons = []
+        const input = []
         for (let i = 0, j = 0; i < contents.length; i++, j++) {
-            const path = files[j]
-            const stageContent = contents[i].data
-            const liveContent = contents[++i].data
-            if (stageContent === liveContent) {
-                continue
-            }
-
-            if (path.endsWith(consts.JSON) && typeof stageContent === 'object' && typeof liveContent === 'object') {
-                if (!isEqual(stageContent, liveContent)) {
-                    changedJsons.push(path)
-                }
-            } else {
-                changedFiles.push(path)
-            }
+            input.push({
+                path: files[j],
+                stage: contents[i].data,
+                live: contents[++i].data
+            })
         }
 
-        console.log(`The following ${changedFiles.length} files differ when comparing stage version to live version (note that these might be false positives for 'html' as the comparison method is very crude):`)
-        consts.printResources(changedFiles)
-
-        console.log(`The following ${changedJsons.length} JSON files differ when comparing stage version to live version (deep compare was used):`)
-        consts.printResources(changedJsons)
+        comparer.compare(input)
     }).catch(console.log)
-}
-
-function isEqual(value, other) {
-    if (lodash.isEqualWith(value, other)) {
-        return true
-    }
-
-    const cleanValue = clearObject(value)
-    const cleanOther = clearObject(other)
-    return lodash.isEqualWith(cleanValue, cleanOther)
-}
-
-function clearObject(obj) {
-    const result = { }
-    for (const key in obj) {
-        if (SKIP_PROPS.includes(key)) {
-            continue
-        }
-
-        const value = obj[key]
-        const type = typeof value
-        if (type == 'string') {
-            result[key] = value
-        } else {
-            result[key] = clearObject(value)
-        }
-    }
-
-    return result
 }
